// src/types/language.ts
export type Language = 'en' | 'tr';

export interface Translations {
  appTitle: string;
  difficulty: string;
  easy: string;
  medium: string;
  hard: string;
  mixed: string;
  enableTwoDigitOperations: string;
  enableTimer: string;
  time: string;
  enterAnswer: string;
  submit: string;
  nextQuestion: string;
  correct: string;
  incorrect: string;
  timeIsUp: string;
  theCorrectAnswerIs: string;
}